requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
pytest
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart